requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
httpx==0.28.1
google-genai==0.8.0
python-dotenv==1.0.1
chromadb==0.5.20
sentence-transformers==3.3.1
pypdf==5.1.0
tiktoken==0.10.0
streamlit>=1.28.0
elevenlabs>=1.0.0
pydub>=0.25.1

